unify ob `limit` handling
In class definition:
ob = {
	'limits': [], # put limits available via websocket here
	'fetch_limit': 500, # the limit used to fetch snapshots
}
In case subscription's limit doesn't match any of value of the 'limits',
it is rounded up to the closest 'limit'. If not upper limit exists, it
is either a) replaced with `None` b) rounded down if 'limit'
corresponding to `None` (exchange's default limit) would be smaller

If no 'fetch_limit' is defined, the (resolved) 'limit' specified in
subscription is used.